yappcap.pyx: Undo UTF-8-ification of strings; stick with bytes
In commit 048e5ee502360dd81b6ffb6b15b7f3cbc3922829, C strings were
converted to UTF-8 to assign them to Python strings. While this works
just fine, this makes using the strings much more onerous for users
of the library.

This patch removes those changes, and instead declares the Python
strings as bytes. This allows for direct assignment to the C strings,
and allows for caller of the library to manipulate the strings as they
would expect.